rename: rename file name Pascal Case to kebab-case
